backend/requirements.txt
Flask==2.0.1
pytesseract==0.3.8
Pillow==8.2.0
google-api-python-client==2.15.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-cloud-vision==2.4.2
pymongo==3.11.4
